pyproject.toml
[project]
name = "Funcall" # Must be a unique name on PyPI
version = "0.0.1"
description = "A script to use Function Gemma capabilities."
authors = [
    { name = "Ali Rezagholizadeh", email = "alirezagholizadeh70@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
#[tool.setuptools.dynamic]
dependencies = [
    "python-dotenv",
    "torch",
    "huggingface_hub",
    "transformers",
    "accelerate",
    "tzdata"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project.urls]
Homepage = "https://github.com/AliRezagholizadeh/Funcall#"
Issues = "https://github.com/AliRezagholizadeh/Funcall/issues"
